+++ nix/modules/home-manager/openclaw/files.nix
@@ -197,9 +197,7 @@ let
     && !(lib.hasInfix "\t" name)
     && !(lib.hasInfix "\n" name)
     && name != ".."
-    && lib.all (segment: segment != "" && segment != "." && segment != "..") (
-      lib.splitString "/" name
-    );
+    && lib.all (segment: segment != "" && segment != "." && segment != "..") (lib.splitString "/" name);
 
   invalidWorkspaceFileNames = lib.filter (name: !(validWorkspaceFileName name)) (
     lib.attrNames extraWorkspaceFiles